Descr -> Concern Model, small fix, add migrate for remove descr
