+++ .rake_tasks~
@@ -1,0 +1,66 @@
+about
+assets:clean[keep]
+assets:clobber
+assets:environment
+assets:precompile
+cache_digests:dependencies
+cache_digests:nested_dependencies
+db:create
+db:drop
+db:fixtures:load
+db:migrate
+db:migrate:status
+db:rollback
+db:schema:cache:clear
+db:schema:cache:dump
+db:schema:dump
+db:schema:load
+db:seed
+db:setup
+db:structure:dump
+db:structure:load
+db:version
+diagram:all
+diagram:all_with_engines
+diagram:controllers:all
+diagram:controllers:brief
+diagram:controllers:brief_with_engines
+diagram:controllers:complete
+diagram:controllers:complete_with_engines
+diagram:models:all
+diagram:models:brief
+diagram:models:brief_with_engines
+diagram:models:complete
+diagram:models:complete_with_engines
+diagram:setup:create_new_doc_folder_if_needed
+doc:app
+log:clear
+middleware
+notes
+notes:custom
+rails:template
+rails:update
+refinery:override
+refinery:override:list
+refinery:uncrudify
+refinery_images:install:migrations
+refinery_inquiries:install:migrations
+refinery_pages:install:migrations
+refinery_resources:install:migrations
+refinery_settings:install:migrations
+routes
+secret
+seo_meta:install:migrations
+stats
+test
+test:all
+test:all:db
+test:db
+time:zones:all
+tmp:clear
+tmp:create
+translate:changed
+translate:google
+translate:lost_in_translation
+translate:lost_in_translation_all
+translate:merge_keys
