+++ app/models/ws_param.rb
@@ -1,4 +1,5 @@
 class WsParam < ActiveRecord::Base
+  include DescriptionsModule
   nilify_blanks
   belongs_to :user
   has_many :descriptions, as: :rec, dependent: :destroy
