+++ app/models/ws_set_model_run.rb
@@ -1,4 +1,5 @@
 class WsSetModelRun < ActiveRecord::Base
+  include DescriptionsModule
   nilify_blanks
   has_many :descriptions, as: :rec, dependent: :destroy
   has_many :ws_model_runs_set_model_runs, :dependent => :destroy
