+++ config/application.rb
@@ -13,7 +13,7 @@ module WsDss
     # -- all .rb files in that directory are automatically loaded.
     
     # Autoload lib/ folder including all subdirectories
-    config.autoload_paths += Dir["#{config.root}/lib/**/"]
+    #config.autoload_paths += Dir["#{config.root}/lib/**/"]
     
     # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
     # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
