+++ db/migrate/20170503042804_remove_descr_from_ws_models.rb
@@ -1,0 +1,5 @@
+class RemoveDescrFromWsModels < ActiveRecord::Migration
+  def change
+    remove_column :ws_models, :descr, :string
+  end
+end
