+++ db/migrate/20170503042842_remove_descr_from_ws_params.rb
@@ -1,0 +1,5 @@
+class RemoveDescrFromWsParams < ActiveRecord::Migration
+  def change
+    remove_column :ws_params, :descr, :string
+  end
+end
