+++ db/migrate/20170503042910_remove_descr_from_ws_set_model_runs.rb
@@ -1,0 +1,5 @@
+class RemoveDescrFromWsSetModelRuns < ActiveRecord::Migration
+  def change
+    remove_column :ws_set_model_runs, :descr, :string
+  end
+end
